requirements.txt
blinker==1.9.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.1.8
decorator==5.2.1
Flask==3.1.3
flask-cors==6.0.2
gTTS==2.5.4
gunicorn==25.3.0
idna==3.13
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
moviepy==2.2.1
numpy==2.4.4
packaging==26.2
pillow==11.3.0
proglog==0.1.12
python-dotenv==1.2.2
requests==2.33.1
tqdm==4.67.3
urllib3==2.6.3
Werkzeug==3.1.8
yt-dlp==2026.3.17
